LPS-62205 Move the code to receive a notification of changing a configuration to another component so avoid the portlet to be redeployed
